Update funding.json
Feedback from FLOSS

Wellknown URL is required if the entity webpage URL and the URL of the funding.json manifest do not have the same hostname. This url should have the same hostname as the url to verify that the publisher of the manifest is authorised to solicit funding on behalf of the url. In your case the entity URL is on https://www.hotosm.org and the funding.json is on github. You should have something like `https://www.hotosm.org/*/.well-known/funding-manifest-urls`